requirements.txt
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.2
matplotlib==3.5.2
MouseInfo==0.1.3
mss==6.1.0
numpy==1.22.4
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
PyAutoGUI==0.9.53
PyBrain==0.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
six==1.16.0
wheel==0.37.1
